added more validation checks
